+++ include/battery_utils.h
@@ -25,6 +25,8 @@
 namespace BATTERY_Utils {
 
     String  getPercentVoltageBattery(float voltage);
+
+    float   readBatteryVoltage();
     void    monitor();
     
 }
